+++ examples/make-fuzzer-input.cpp
@@ -24,7 +24,7 @@
 #include <vector>
 
 #include <deepstream/buffer.hpp>
-#include <deepstream/message_builder.hpp>
+#include "../src/message_builder.hpp"
 #include "../src/random.hpp"
 
 #include <cassert>
