+++ src/event.cpp
@@ -20,7 +20,7 @@
 
 #include <deepstream/buffer.hpp>
 #include <deepstream/event.hpp>
-#include <deepstream/message_builder.hpp>
+#include "message_builder.hpp"
 
 #include <cassert>
