+++ src/message_proxy.cpp
@@ -16,7 +16,7 @@
 #include <ostream>
 
 #include <deepstream/buffer.hpp>
-#include <deepstream/message_proxy.hpp>
+#include "message_proxy.hpp"
 
 #include <cassert>
