+++ src/message_proxy.hpp
@@ -21,7 +21,7 @@
 #include <iosfwd>
 #include <vector>
 
-#include <deepstream/message.hpp>
+#include "message.hpp"
 
 namespace deepstream {
 struct Buffer;
