+++ src/presence.cpp
@@ -19,7 +19,7 @@
 #include <stdexcept>
 
 #include <deepstream/buffer.hpp>
-#include <deepstream/message_builder.hpp>
+#include "message_builder.hpp"
 #include <deepstream/presence.hpp>
 
 #include <cassert>
