+++ src/random.cpp
@@ -17,7 +17,7 @@
 #include <limits>
 
 #include <deepstream/buffer.hpp>
-#include <deepstream/message_builder.hpp>
+#include "message_builder.hpp"
 #include "random.hpp"
 
 #include <cassert>
