+++ test/message_builder.cpp
@@ -24,7 +24,7 @@
 #include <stdexcept>
 
 #include <deepstream/buffer.hpp>
-#include <deepstream/message_builder.hpp>
+#include "../src/message_builder.hpp"
 
 namespace deepstream {
