+++ test/parser.cpp
@@ -25,8 +25,8 @@
 #include <vector>
 
 #include <deepstream/buffer.hpp>
-#include <deepstream/message.hpp>
-#include <deepstream/message_builder.hpp>
+#include "../src/message.hpp"
+#include "../src/message_builder.hpp"
 #include "../src/parser.hpp"
 #include "../src/random.hpp"
 #include "../src/scope_guard.hpp"
